Ensure ActionCable is thread-safe
Although this is unlikely to cause any problems specially because this
is first run as "run ActionCable.server" and it's very unlikely to call
multiple rack "run" methods concurrently and even if we did so it's not
guaranteed that would cause any troubles, it's much better not having to
worry about this and just ensuring this is thread-safe.

It might seem that @a ||= A.new pattern runs atomically, but this is not
guaranteed nor even by MRI. Here's an example:

class A
def initialize
    sleep 1
end
def self.instance
    @@instance ||= new
end
end

instances = [] # I know this is not thread-safe but it demonstrates the
               # problem anyway

(1..2).map{Thread.start{instances << A.instance}}.each &:join
instances[0] == instances[1] # false

Using the singleton pattern makes the intention clearer and worry-free.